simplify handling of default values for properties
Instead of asking for a "function that returns a default" as part of the
property definition, we just expect default to be set in the widget's
constructor. The default value is almost always the one that the widget
is constructed with anyway.

For some some reason, this also fixes #75. I'm not sure why. But I'll take it.